Loading Scene, Coroutines, Spinner
A loading scene is added. Loader script reads the input file with a coroutine. This coroutine creates a thread to read the file. During read, a spinning circle is diplayed to the user. After the reading is done, transition to the main scene happens. Main scene has an opening transition. Plane labels are added. (XY, YZ, XZ planes)